Rename file CardMetadata to ICardMetadata.
